symmetric-asymmetric/requirements.txt
boto3==1.13.12
cryptography==2.9.2
pycrypto==2.6.1
